use np.float

+++ tests/test_metrics.py
@@ -69,7 +69,7 @@ def legacy_mean_fscore(results,
                        beta=1):
     num_imgs = len(results)
     assert len(gt_seg_maps) == num_imgs
-    total_mat = np.zeros((num_classes, num_classes), dtype=np.float)
+    total_mat = np.zeros((num_classes, num_classes), dtype=np.float32)
     for i in range(num_imgs):
         mat = get_confusion_matrix(
             results[i], gt_seg_maps[i], num_classes, ignore_index=ignore_index)
